refactor(menu): extract CSV parser and create reusable bulk upload component
